requirements.txt
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
pydantic>=2.5.0
markdown2>=2.4.0
xhtml2pdf>=0.2.17
python-docx>=1.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
websockets>=12.0
celery>=5.3.4
redis>=5.0.0
